Add support for having Drupal bootstrapped from other than sites/default.
This adds support for multisite projects. build.sh adds a parameter to drush, -l,
that if set will make build.sh new and build.sh update work for legacy projects
that are based on sites.php/have the settings file in a different place than the
normal location - sites/default.